requirements.txt
pytest
pytest-logger
pytest-html
pytest-json

pyhamcrest
jsonschema

requests
slackclient
